refactor(coin-evm): remove cyclic imports
