+++ .changeset/clean-shirts-call.md
@@ -1,0 +1,9 @@
+---
+"@ledgerhq/coin-evm": minor
+"@ledgerhq/live-common": minor
+"@ledgerhq/coin-tester-evm": minor
+"ledger-live-desktop": minor
+"live-mobile": minor
+---
+
+refactor(coin-evm): remove cyclic imports
