+++ apps/ledger-live-desktop/src/renderer/families/evm/SendAmountFields/SelectFeeStrategy.tsx
@@ -1,5 +1,5 @@
 import { isStrategyDisabled } from "@ledgerhq/coin-evm/editTransaction/index";
-import { getEstimatedFees } from "@ledgerhq/coin-evm/logic";
+import { getEstimatedFees } from "@ledgerhq/coin-evm/utils";
 import { getTypedTransaction } from "@ledgerhq/coin-evm/transaction";
 import {
   Transaction as EvmTransaction,
