+++ apps/ledger-live-mobile/src/families/evm/EvmCustomFees/Evm1559CustomFees.tsx
@@ -1,7 +1,7 @@
 import { getAccountBridge } from "@ledgerhq/live-common/bridge/index";
 import useBridgeTransaction from "@ledgerhq/live-common/bridge/useBridgeTransaction";
 import { formatCurrencyUnit } from "@ledgerhq/live-common/currencies/index";
-import { getGasLimit } from "@ledgerhq/coin-evm/logic";
+import { getGasLimit } from "@ledgerhq/coin-evm/utils";
 import type { EvmTransactionEIP1559 } from "@ledgerhq/coin-evm/types/index";
 import { Account } from "@ledgerhq/types-live";
 import { useTheme } from "@react-navigation/native";
