+++ apps/ledger-live-mobile/src/families/evm/SelectFeesStrategy.tsx
@@ -1,5 +1,5 @@
 import { isStrategyDisabled } from "@ledgerhq/coin-evm/editTransaction/index";
-import { getEstimatedFees } from "@ledgerhq/coin-evm/logic";
+import { getEstimatedFees } from "@ledgerhq/coin-evm/utils";
 import { getTypedTransaction } from "@ledgerhq/coin-evm/transaction";
 import type { FeeData, GasOptions, Strategy, Transaction } from "@ledgerhq/coin-evm/types/index";
 import { getFeesCurrency, getFeesUnit, getMainAccount } from "@ledgerhq/live-common/account/index";
