+++ libs/coin-modules/coin-evm/package.json
