+++ libs/coin-modules/coin-evm/src/__tests__/unit/synchronization.unit.test.ts
@@ -640,7 +640,6 @@ describe("EVM Family", () => {
               },
             };
           });
-          console.log(etherscanAPI?.default.getLastOperations);
           const spy = jest.spyOn(etherscanAPI, "getLastERC1155Operations");
 
           await synchronization.getAccountShape(
