+++ libs/coin-modules/coin-evm/src/buildOptimisticOperation.ts
@@ -9,8 +9,9 @@ import { encodeNftId } from "@ledgerhq/coin-framework/nft/nftId";
 import { encodeOperationId } from "@ledgerhq/coin-framework/operation";
 import { findSubAccountById } from "@ledgerhq/coin-framework/account/index";
 import { EvmNftTransaction, Transaction as EvmTransaction } from "./types";
-import { getEstimatedFees, isNftTransaction } from "./logic";
+import { isNftTransaction } from "./logic";
 import { toTransactionRaw } from "./transaction";
+import { getEstimatedFees } from "./utils";
 
 /**
  * Build an optimistic operation for the coin of the integration (e.g. Ether for Ethereum)
