+++ libs/coin-modules/coin-evm/src/createTransaction.ts
@@ -1,13 +1,7 @@
 import BigNumber from "bignumber.js";
 import { AccountBridge, AccountLike } from "@ledgerhq/types-live";
 import { Transaction as EvmTransaction } from "./types";
-import { DEFAULT_GAS_LIMIT } from "./transaction";
-
-/**
- * arbitrary default value for a new transaction nonce
- * the actual nonce will be set by `prepareForSignOperation`
- */
-export const DEFAULT_NONCE = -1;
+import { DEFAULT_GAS_LIMIT, DEFAULT_NONCE } from "./utils";
 
 const getChainId = (account: AccountLike): number => {
   if (account.type === "Account") {
