+++ libs/coin-modules/coin-evm/src/editTransaction/hasMinimumFunds.ts
@@ -1,8 +1,8 @@
 import { getEnv } from "@ledgerhq/live-env";
 import type { Account, AccountLike } from "@ledgerhq/types-live";
 import { BigNumber } from "bignumber.js";
-import { getEstimatedFees } from "../logic";
 import type { Transaction } from "../types/index";
+import { getEstimatedFees } from "../utils";
 
 export const hasMinimumFundsToCancel = ({
   mainAccount,
