+++ libs/coin-modules/coin-evm/src/getTransactionStatus.ts
@@ -25,9 +25,8 @@ import {
   TransactionStatusCommon,
 } from "@ledgerhq/types-live";
 import { getMinEip1559Fees, getMinLegacyFees } from "./editTransaction/getMinEditTransactionFees";
-import { eip1559TransactionHasFees, getEstimatedFees, legacyTransactionHasFees } from "./logic";
+import { eip1559TransactionHasFees, legacyTransactionHasFees } from "./logic";
 import { NotEnoughNftOwned, NotOwnedNft, QuantityNeedsToBePositive } from "./errors";
-import { DEFAULT_GAS_LIMIT } from "./transaction";
 import {
   EditType,
   Transaction as EvmTransaction,
@@ -35,6 +34,7 @@ import {
   EvmTransactionLegacy,
   TransactionStatus,
 } from "./types";
+import { DEFAULT_GAS_LIMIT, getEstimatedFees } from "./utils";
 
 type ValidatedTransactionFields =
   | "recipient"
